Visualizing the difference of nighttime lights
Here, I visualized the nighttime lights change of the Philippines during December 2019, 2020, and 2021 which is the peak Christmas season of the nation. Upon visual analysis, there has been a slight decrease from December 2019 to December 2020 and an increase from December 2020 to December 2021. This may be due to the restrictions imposed during the said year.